LayoutTests/inspector/canvas/updateShader-webgpu-expected.txt

WebGPU tests for Canvas.updateShader command.


== Running test suite: Canvas.updateShader.WebGPU
-- Running test case: Canvas.updateShader.WebGPU.Compute.Valid
PASS: Source should have changed.

-- Running test case: Canvas.updateShader.WebGPU.Compute.Invalid
PASS: Should produce an exception.
Error: Failed to update shader of given shaderType for given programId
PASS: Source should have changed.

-- Running test case: Canvas.updateShader.WebGPU.Vertex.Valid
PASS: Source should have changed.

-- Running test case: Canvas.updateShader.WebGPU.Vertex.Invalid
PASS: Should produce an exception.
Error: Failed to update shader of given shaderType for given programId
PASS: Source should have changed.

-- Running test case: Canvas.updateShader.WebGPU.Fragment.Valid
PASS: Source should have changed.

-- Running test case: Canvas.updateShader.WebGPU.Fragment.Invalid
PASS: Should produce an exception.
Error: Failed to update shader of given shaderType for given programId
PASS: Source should have changed.
